Skip tournaments with no data
Things exploded otherwise, too - since I take the tournament type from the first row of results, and that won't work if there are no results

+++ app/models/lists_juggler_importer.rb
@@ -61,29 +61,31 @@ class ListsJugglerImporter
   def process_data(tournament_id, tournament_data)
     ActiveRecord::Base.transaction do
       begin
-        tournament_type = TournamentType.find_by!(name: tournament_data[1][1])
-        tournament      = Tournament.find_by!(tournament_type:  tournament_type,
-                                              lists_juggler_id: tournament_id)
-        tournament.squadrons.destroy_all
-        header_row = tournament_data[0]
-        tournament_data[1..-1].each do |ship_data|
-          if ship_data[9].present? && ship_data[10].present? # ignore all rows without ship and/or pilot
-            player             = Player.find_or_create_by!(name: ship_data[3])
-            faction            = Faction.find_or_create_by!(name: ship_data[4])
-            squadron           = Squadron.find_or_create_by!(faction:          faction,
-                                                             tournament:       tournament,
-                                                             player:           player,
-                                                             lists_juggler_id: ship_data[8])
-            ship               = Ship.find_or_create_by!(name: ship_data[9])
-            pilot              = Pilot.find_or_create_by!(faction: faction,
-                                                          ship:    ship,
-                                                          name:    ship_data[10])
-            ship_configuration = ShipConfiguration.create!(squadron: squadron, pilot: pilot)
-            ship_data[11..-1].each.with_index do |upgrade_name, index|
-              if upgrade_name.present?
-                upgrade_type = UpgradeType.find_or_create_by!(name: header_row[11 + index].split('.')[0])
-                upgrade      = Upgrade.find_or_create_by!(upgrade_type: upgrade_type, name: upgrade_name)
-                ship_configuration.upgrades << upgrade
+        if tournament_data.length > 1
+          tournament_type = TournamentType.find_by!(name: tournament_data[1][1])
+          tournament      = Tournament.find_by!(tournament_type:  tournament_type,
+                                                lists_juggler_id: tournament_id)
+          tournament.squadrons.destroy_all
+          header_row = tournament_data[0]
+          tournament_data[1..-1].each do |ship_data|
+            if ship_data[9].present? && ship_data[10].present? # ignore all rows without ship and/or pilot
+              player             = Player.find_or_create_by!(name: ship_data[3])
+              faction            = Faction.find_or_create_by!(name: ship_data[4])
+              squadron           = Squadron.find_or_create_by!(faction:          faction,
+                                                               tournament:       tournament,
+                                                               player:           player,
+                                                               lists_juggler_id: ship_data[8])
+              ship               = Ship.find_or_create_by!(name: ship_data[9])
+              pilot              = Pilot.find_or_create_by!(faction: faction,
+                                                            ship:    ship,
+                                                            name:    ship_data[10])
+              ship_configuration = ShipConfiguration.create!(squadron: squadron, pilot: pilot)
+              ship_data[11..-1].each.with_index do |upgrade_name, index|
+                if upgrade_name.present?
+                  upgrade_type = UpgradeType.find_or_create_by!(name: header_row[11 + index].split('.')[0])
+                  upgrade      = Upgrade.find_or_create_by!(upgrade_type: upgrade_type, name: upgrade_name)
+                  ship_configuration.upgrades << upgrade
+                end
               end
             end
           end
